API/C#: Update Dto and Fix Mapper Profile for unmatching profiles

+++ CarDB-Csharp-API/Models/Entities/Car.cs
@@ -19,11 +19,11 @@ namespace CarDB_Csharp_API.Models.Entities
         public required string Origin {get; set;}
         [Required]
         public required int ModelYear {get; set;}
-        public int? Acceleration {get; set;}
-        public int? HorsePower {get; set;}
-        public int? MPG {get; set;}
-        public int? Cylinders {get; set;}
-        public int? Weight {get; set;}
-        public int? Displacement {get; set;}
+        public float? Acceleration {get; set;}
+        public float? HorsePower {get; set;}
+        public float? MPG {get; set;}
+        public float? Cylinders {get; set;}
+        public float? Weight {get; set;}
+        public float? Displacement {get; set;}
     }
 }
